Added feature: Image as part of HTML post content
Added drag/drop, file picker for images.
Each image is sent by a POST request to /upload before form submission. A post request handler accepts these, and stores each file in public/uploads/postFiles/ directory, and returns its path on the server side in a JSON as response. Then the TinyMCE functionality 'replaces' the src attributes of each image with the new path, on the server, so the post content will be completely visible to all readers.