Edit: Tabs on Profile page.
It's still static, haven't made an EJS file.

Made a custom element of user-post.
Note: The element user-post is a limited view of the post.
It shows the title, time and date (formatted), an excerpt of the entire content, the subforum it was posted in, upvotes and downvotes (with buttons), along with the share buttons.

The tabs are About, Recent Posts, Followed Subforums and Communities.
About tab has a bio, and a list-group of interests.
Recent Posts tab has a list of user-posts
Followed Subforums and Communities tabs have lists.

Functionality of sharing, viewing post, upvoting and downvoting is left.

Problem: How to handle categories in a user-post?