Fix: Chatting for users with an underscore in their usernames
Earlier there was a check in the "joinRoom" event handler in routes/chat.js to verify whether the requesting user was part of the requested room. This check used to split the roomName by underscore. Hence the bug.
It is in comments for now.

Also fixed the link for chatWithUser on the chat page.